+++ sdk/cosmos/azure_data_cosmos/src/clients/cosmos_client.rs
@@ -4,7 +4,7 @@
 use crate::{
     clients::DatabaseClient,
     models::DatabaseProperties,
-    pipeline::{AuthorizationPolicy, CosmosPipeline},
+    pipeline::{AuthorizationPolicy, GatewayPipeline},
     resource_context::{ResourceLink, ResourceType},
     CosmosClientOptions, CreateDatabaseOptions, FeedPager, Query, QueryDatabasesOptions,
 };
@@ -15,18 +15,20 @@ use azure_core::{
 use serde::Serialize;
 use std::sync::Arc;
 
+use crate::constants::COSMOS_ALLOWED_HEADERS;
 use crate::cosmos_request::CosmosRequest;
 use crate::operation_context::OperationType;
 use crate::routing::global_endpoint_manager::GlobalEndpointManager;
 #[cfg(feature = "key_auth")]
 use azure_core::credentials::Secret;
-use azure_core::http::RetryOptions;
+use azure_core::http::{LoggingOptions, RetryOptions};
 
 /// Client for Azure Cosmos DB.
 #[derive(Debug, Clone)]
 pub struct CosmosClient {
     databases_link: ResourceLink,
-    pipeline: Arc<CosmosPipeline>,
+    pipeline: Arc<GatewayPipeline>,
+    global_endpoint_manager: GlobalEndpointManager,
 }
 
 impl CosmosClient {
@@ -53,9 +55,16 @@ impl CosmosClient {
         options: Option<CosmosClientOptions>,
     ) -> azure_core::Result<Self> {
         let options = options.unwrap_or_default();
-        let mut client_options = options.client_options;
+        let endpoint: Url = endpoint.parse()?;
+        let mut client_options = options.client_options.clone();
         client_options.retry = RetryOptions::none();
-
+        client_options.logging = LoggingOptions {
+            additional_allowed_header_names: COSMOS_ALLOWED_HEADERS
+                .iter()
+                .map(|h| std::borrow::Cow::Borrowed(h.as_str()))
+                .collect(),
+            additional_allowed_query_params: vec![],
+        };
         let pipeline_core = azure_core::http::Pipeline::new(
             option_env!("CARGO_PKG_NAME"),
             option_env!("CARGO_PKG_VERSION"),
@@ -67,21 +76,21 @@ impl CosmosClient {
             None,
         );
 
-        let global_endpoint_manager = GlobalEndpointManager::new(
-            endpoint.parse()?,
-            options.application_preferred_regions,
-            pipeline_core.clone(),
-        );
+        let preferred_regions = options.application_preferred_regions.clone();
+        let global_endpoint_manager =
+            GlobalEndpointManager::new(endpoint.clone(), preferred_regions, pipeline_core.clone());
 
-        let pipeline = Arc::new(CosmosPipeline::new(
-            endpoint.parse()?,
+        let pipeline = Arc::new(GatewayPipeline::new(
+            endpoint,
             pipeline_core,
-            global_endpoint_manager,
+            global_endpoint_manager.clone(),
+            options,
         ));
 
         Ok(Self {
             databases_link: ResourceLink::root(ResourceType::Databases),
             pipeline,
+            global_endpoint_manager,
         })
     }
 
@@ -108,8 +117,17 @@ impl CosmosClient {
         options: Option<CosmosClientOptions>,
     ) -> azure_core::Result<Self> {
         let options = options.unwrap_or_default();
-        let mut client_options = options.client_options;
+        let endpoint: Url = endpoint.parse()?;
+
+        let mut client_options = options.client_options.clone();
         client_options.retry = RetryOptions::none();
+        client_options.logging = LoggingOptions {
+            additional_allowed_header_names: COSMOS_ALLOWED_HEADERS
+                .iter()
+                .map(|h| std::borrow::Cow::Borrowed(h.as_str()))
+                .collect(),
+            additional_allowed_query_params: vec![],
+        };
 
         let pipeline_core = azure_core::http::Pipeline::new(
             option_env!("CARGO_PKG_NAME"),
@@ -120,21 +138,21 @@ impl CosmosClient {
             None,
         );
 
-        let global_endpoint_manager = GlobalEndpointManager::new(
-            endpoint.parse()?,
-            options.application_preferred_regions,
-            pipeline_core.clone(),
-        );
+        let preferred_regions = options.application_preferred_regions.clone();
+        let global_endpoint_manager =
+            GlobalEndpointManager::new(endpoint.clone(), preferred_regions, pipeline_core.clone());
 
-        let pipeline = Arc::new(CosmosPipeline::new(
-            endpoint.parse()?,
+        let pipeline = Arc::new(GatewayPipeline::new(
+            endpoint,
             pipeline_core,
             global_endpoint_manager.clone(),
+            options,
         ));
 
         Ok(Self {
             databases_link: ResourceLink::root(ResourceType::Databases),
             pipeline,
+            global_endpoint_manager,
         })
     }
 
@@ -173,7 +191,11 @@ impl CosmosClient {
     /// # Arguments
     /// * `id` - The ID of the database.
     pub fn database_client(&self, id: &str) -> DatabaseClient {
-        DatabaseClient::new(self.pipeline.clone(), id)
+        DatabaseClient::new(
+            self.pipeline.clone(),
+            id,
+            self.global_endpoint_manager.clone(),
+        )
     }
 
     /// Gets the endpoint of the database account this client is connected to.
@@ -244,7 +266,7 @@ impl CosmosClient {
 
         let cosmos_request =
             CosmosRequest::builder(OperationType::Create, self.databases_link.clone())
-                .headers(&options.throughput)
+                .request_headers(&options.throughput)
                 .json(&RequestBody { id })
                 .build()?;